requirements.txt
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
groq>=0.9.0
python-dotenv>=1.0.0
